ci(win): silence two more clang-cl warnings
  - lib/CMakeLists.txt: extend the source-level
    -Wno-unused-but-set-variable / -Wno-unused-private-field path to
    clang-cl (CMAKE_CXX_COMPILER_ID == 'Clang' even when MSVC=1). Was
    gated `if(NOT MSVC)`, so clang-cl on Windows kept tripping on
    bison's yynerrs_ generated 'set but not used' variable. macOS Clang
    and Linux GCC keep going through the NOT MSVC branch unchanged.

  - test_expression_operators.cpp: SignedUnsignedComparison test was
    storing 0xFFFFFFFFFFFFFFFF in an `unsigned long` literal that's
    32-bit on Windows (LLP64). Switch to `unsigned long long` so the
    semantics match across LLP64 (Windows) and LP64 (Linux/macOS).

+++ lib/test/passes/transformations/test_expression_operators.cpp
@@ -1619,7 +1619,7 @@ TEST(MiscBignum, SignedUnsignedComparison)
 {
     const mpz_class neg(-1);
     const unsigned int u_small = 1u;
-    const unsigned long u_huge = 0xFFFFFFFFFFFFFFFFull;
+    const unsigned long long u_huge = 0xFFFFFFFFFFFFFFFFull;
 
     // A negative bignum is less than any unsigned, no matter how large.
     ASSERT_TRUE(neg < u_small);
